Took away databse wipe

+++ src/edu/gatech/mdiamond8/partytrack/Main.java
@@ -15,10 +15,14 @@ import java.sql.Statement;
 
 public class Main extends Application {
     public static void main(String[] args) throws IOException {
-        try {
-            NetworkCode.deleteAttendees();
-        } catch (Exception ex) {
+        if (args.length == 1) {
+            if (args[0].equals("Delete")) {
+                try {
+                    NetworkCode.deleteAttendees();
+                } catch (Exception ex) {
 
+                }
+            }
         }
         try {
             NetworkCode.makeAttendees();
